requirements.txt
Django>=3.2
djangorestframework>=3.13.1